Adding text to clarify dummy and system toolchains and forward references
The table contained a row for the `dummy` toochain.  Since it will still be
seen by people using versions prior to v4.0, the entry in the table is being
left but text about deprecation is added in the body of the table, and some
additional text is included in the note, along with references to other
pages.